Clean out create-react-app boilerplate
